+++ src/asyncSensor.js
@@ -1,4 +1,4 @@
-jasmineui.define('asyncSensor', ['globals', 'logger', 'loadListener'], function (globals, logger, loadListener) {
+jasmineui.define('client?asyncSensor', ['globals', 'logger', 'loadListener'], function (globals, logger, loadListener) {
     var window = globals.window;
     window.jasmineui = window.jasmineui || {};
     var asyncSensors = window.jasmineui.asyncSensors = window.jasmineui.asyncSensors || {};