+++ src/config.js
@@ -1,12 +1,27 @@
 jasmineui.define('config', ['globals', 'persistentData'], function (globals, persistentData) {
-    var popupMode = !!globals.jstestdriver;
     var pd = persistentData();
-    var clientMode = pd.specs && pd.specIndex < pd.specs.length;
 
-    return {
+    var config = {
         logEnabled:false,
         waitsForAsyncTimeout:5000,
-        popupMode:popupMode,
-        clientMode:clientMode
+        loadMode:'inplace',
+        closeTestWindow:true
+    };
+
+    function merge(obj) {
+        var prop;
+        for (prop in obj) {
+            config[prop] = obj[prop];
+        }
     }
+
+    if (pd.config) {
+        merge(pd.config);
+    }
+    if (globals.jasmineuiConfig) {
+        merge(globals.jasmineuiConfig);
+    }
+    pd.config = config;
+
+    return config;
 });