+++ src/main.js
@@ -1,35 +1,3 @@
 (function () {
-
-
-
-    jasmineui.require(['simulateEvent'], function (simulate) {
-        window.simulate = simulate;
-    });
-
-    var config;
-    jasmineui.require(['config'], function(_config) {
-        config = _config;
-    });
-    if (config.clientMode) {
-        jasmineui.require(['describeUiClient'], function (describeUi) {
-            window.xdescribeUi = window.xdescribe;
-            window.describe = describeUi.describe;
-            window.describeUi = describeUi.describeUi;
-            window.beforeLoad = describeUi.beforeLoad;
-            window.runs = describeUi.runs;
-            window.waitsFor = describeUi.waitsFor;
-            window.waits = describeUi.waits;
-            jasmineui.inject = describeUi.inject;
-        });
-    } else {
-        jasmineui.require(['describeUiServer'], function (describeUi) {
-            window.describeUi = describeUi.describeUi;
-            window.it = describeUi.it;
-            window.beforeLoad = describeUi.beforeLoad;
-            window.describe = describeUi.describe;
-            window.xdescribeUi = window.xdescribe;
-            jasmineui.inject = describeUi.inject;
-
-        });
-    }
+    jasmineui.require.all();
 })();