+++ test/lib/jstd-jasmineui-cfg.js
@@ -1,0 +1,4 @@
+jasmineuiConfig = {
+    loadMode: 'popup',
+    closeTestWindow: true
+};