+++ test/ui/inject/sampleInjectedCallback.js
@@ -1,7 +1,0 @@
-var sampleInjectedCallbackCalled = false;
-
-jasmineui.inject(function() {
-    sampleInjectedCallbackCalled = true;
-});
-
-jasmineui.inject("sampleInjectedScript.js");